requirements.txt
# Core framework
Django==4.2.1
djangorestframework==3.14.0

# Authentication & RBAC
djangorestframework-simplejwt==5.2.2

# Database driver
psycopg2-binary==2.9.9

# Environment variable management
python-dotenv==1.0.0

# WSGI server
gunicorn==20.1.0

# CORS support
django-cors-headers==4.3.0

# Filtering & pagination
django-filter==23.5

# Asynchronous tasks
celery==5.2.7
redis==4.5.0

# API documentation (Swagger/OpenAPI)
drf-yasg==1.21.5

# Audit logging
django-auditlog==2.3.0

# Testing
pytest==7.2.2
pytest-django==4.5.2